tests/requirements-test.txt
pandas
numpy
scikit-learn
umap-learn
python-dotenv
pytest
bertopic